Remove row padding
This is a grid architecture decision that is founded on the principle of only giving padding to elements that require it. By removing padding from rows, the row is really doing just its basic job: containing columns. In the future we will investigate using a larger parent container that will hold all of the rows in an App, and we will apply padding to that.